qorgau-city/src/requirements.txt
Django==4.2.6
asgiref==3.7.2
django-cors-headers==4.2.0
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-rest-swagger==2.2.0
django-redis==5.3.0
boto3==1.26.143
minio==7.2.3
django-storages==1.13.2
Pillow==10.1.0
psycopg2==2.9.9
drf_yasg==1.21.7
PyJWT==2.8.0
pytz==2023.3.post1
rest-framework-simplejwt==0.0.2
setuptools==74.1.2
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
python_docx==1.1.0
redis==5.0.1
django-db-logger==0.1.12
gunicorn==21.2.0
uvicorn==0.23.2
drf-spectacular==0.27.2

celery==5.4.0
django-celery-beat==2.7.0

# WebSocket support
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

reportlab